registro-escolar-main/requirements.txt
black
Django==4.2.9
django-admin-list-filter-dropdown
django-debug-toolbar
django-smart-selects
gunicorn
openpyxl
XlsxWriter
